requirements.txt
Django==4.1.7
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
phonenumbers==8.13.8
Pillow==9.5.0
python-dotenv==1.0.0
psycopg2-binary
drf-yasg==1.21.5
django-rest-swagger==2.2.0
gunicorn==20.1.0
django-cors-headers==3.13.0